Move InsertSizeDistribution to separate file
